# api/openapi/auth-service.openapi.yaml
openapi: 3.1.0
info:
  title: Vista Platform — auth-service API
  version: "0.1.0"
  description: |
    Spec-first contract for auth-service (ADR-0001). Eighth (and final
    pre-spec) vertical slice for the queue: covers the three **cross-cutter**
    endpoints every page in any future surface calls on load.

    ## Scope of this slice

    Three endpoints on the v1 path prefix `/api/v1/auth-service`:

      - `GET /auth/me`           — current user + tenant context (web-ui `auth-context.tsx` on every login + bootstrap)
      - `GET /user/permissions`  — string-array of effective permissions for RBAC gating (web-ui `PermissionGate` / `usePermissions()`)
      - `GET /tenant/features`   — resolved feature-flag map plus usage limits (web-ui `useFeatures()` / `useFeature()`)

    Out of scope for this slice (each is a larger surface that needs its own
    slice with the design track's input):

      - Login / logout / register / refresh / forgot-password / reset-password
      - SSO authorize / callback / link / unlink (both tenant and platform)
      - `/auth/me` writes (PUT, change-password), preferences, EULA, onboarding
      - Sessions, connections, avatar uploads
      - Users-admin CRUD, RBAC role + permission management, permission-check POST
      - Billing, tier selection, subscription tier reads
      - Tenant read/update, branding, UI config
      - Platform-admin and impersonation surfaces

...
    fetches on load). The intro above is kept as the original framing.

...

...

...
servers:
...
paths:
...
  /auth/methods:
    post:
      operationId: authMethods
      tags: [auth]
      summary: List available sign-in methods for an email
      description: >
        Public, pre-session lookup the shared login page uses to discover how a
        given email can sign in. Resolves the user by email and returns the
        tenant's enabled SSO providers (plus the always-present password method),
        so the login page can render "Continue with <provider>" and build the
        SSO authorize URL without a second round-trip.


        The flow: the page POSTs the email here, then for an `sso` method
        navigates the browser to
        `/api/v1/auth-service/auth/sso/{provider_name}/authorize?tenant_id={tenant_id}`
        using the returned `tenant_id` + the method's `provider_name`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthMethodsRequest"
      responses:
        "200":
          description: >
            Available methods for the email. Always 200, even for an unknown
            email (returns just the password method, `tenant_id` empty) so the
            endpoint does not reveal whether an account exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMethodsResponse"
        "400":
          $ref: "#/components/responses/LegacyBadRequest"
        "500":
          $ref: "#/components/responses/LegacyServerError"
      x-quirks:
        - "Unauthenticated/public. A missing or non-email `email` returns 400 {\"error\":\"Invalid request format\"}."
        - "`tenant_id` is empty when the email matches no active user. SSO methods are only included for a known user whose tenant has enabled providers."
        - "A `platform_sso` method type can also appear (user signed up via Google/Microsoft at the platform level on a password_only tenant); the login page may ignore it."
...
  /tenant/sso/providers:
    get:
      operationId: listSSOProviders
      x-edition: enterprise
      tags:
        - tenant-sso
      summary: List the tenant's SSO providers
      description: "Admins (tenant_admin / security_admin) get the full config + group-role mappings; other roles get the minimal enabled-only view for the login page. An unauthenticated request returns an empty list."
      responses:
        "200":
          description: The tenant's SSO providers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SSOProvidersResponse"
        "400":
          $ref: "#/components/responses/LegacyBadRequest"
        "500":
          $ref: "#/components/responses/LegacyServerError"
    post:
      operationId: createSSOProvider
      x-edition: enterprise
      tags:
        - tenant-sso
      summary: Create an SSO provider
      description: "Creates an SSO provider for the tenant. provider_type + provider_name are required; OAuth2 types additionally require client_id/client_secret/auth_url/token_url, SAML types require saml_entity_id/saml_sso_url. Client secrets are encrypted at rest. Returns the created provider (secrets not included)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSSOProviderRequest"
      responses:
        "201":
          description: The created provider.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SSOProviderResponse"
        "400":
          $ref: "#/components/responses/LegacyBadRequest"
        "401":
          $ref: "#/components/responses/LegacyUnauthorized"
        "500":
          $ref: "#/components/responses/LegacyServerError"
  /tenant/sso/providers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: SSO provider UUID.
        schema:
          type: string
          format: uuid
    put:
      operationId: updateSSOProvider
      x-edition: enterprise
      tags:
        - tenant-sso
      summary: Update an SSO provider (partial)
      description: "Partial update of a provider owned by the caller's tenant — only provided fields are changed. At least one field (or group_role_mappings) is required (400 otherwise). client_secret / saml_private_key are re-encrypted at rest. Returns the updated provider."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSSOProviderRequest"
      responses:
        "200":
          description: The updated provider.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SSOProviderResponse"
        "400":
          $ref: "#/components/responses/LegacyBadRequest"
        "401":
          $ref: "#/components/responses/LegacyUnauthorized"
        "403":
          $ref: "#/components/responses/LegacyForbidden"
        "404":
          $ref: "#/components/responses/LegacyNotFound"
        "500":
          $ref: "#/components/responses/LegacyServerError"
    delete:
      operationId: deleteSSOProvider
      x-edition: enterprise
      tags:
        - tenant-sso
      summary: Delete an SSO provider
      description: "Deletes a provider owned by the caller's tenant. Returns 400 if the provider is still in use by any user (with a user_count hint)."
      responses:
        "200":
          description: Provider deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/LegacyBadRequest"
        "401":
          $ref: "#/components/responses/LegacyUnauthorized"
        "403":
          $ref: "#/components/responses/LegacyForbidden"
        "404":
          $ref: "#/components/responses/LegacyNotFound"
        "500":
          $ref: "#/components/responses/LegacyServerError"
  /tenant/sso/providers/{id}/test:
    parameters:
      - name: id
        in: path
        required: true
        description: SSO provider UUID.
        schema:
          type: string
          format: uuid
    post:
      operationId: testSSOProvider
      x-edition: enterprise
      tags:
        - tenant-sso
      summary: Test an SSO provider's configuration
      description: "Validates the provider's stored configuration completeness (no live IdP call) and returns a status/message result."
      responses:
        "200":
          description: The test result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SSOTestResult"
        "400":
          $ref: "#/components/responses/LegacyBadRequest"
        "401":
          $ref: "#/components/responses/LegacyUnauthorized"
        "404":
          $ref: "#/components/responses/LegacyNotFound"
        "500":
          $ref: "#/components/responses/LegacyServerError"
...
components:
...
  responses:
    LegacyBadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LegacyError"
    LegacyUnauthorized:
      description: Missing or invalid auth / tenant context.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LegacyError"
    LegacyNotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LegacyError"
    LegacyForbidden:
      description: Caller lacks the required role/permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LegacyError"
    LegacyServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LegacyError"
  schemas:
...
    AuthMethodsRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    AuthMethod:
      type: object
      description: >
        One available sign-in method. `password` is always present for a known
        user; `sso` entries carry the provider identity needed to start the
        authorize flow. A `platform_sso` type may also appear.
      required: [type, enabled]
      properties:
        type:
          type: string
          enum: [password, sso, platform_sso]
        enabled: { type: boolean }
        provider:
          type: string
          description: "Provider type (e.g. google, microsoft). Present for sso / platform_sso."
        provider_id:
          type: string
          format: uuid
          description: "sso_providers row id. Present for tenant sso methods."
        provider_name:
          type: string
          description: "Tenant-chosen provider name; the URL segment for the authorize endpoint. Present for tenant sso methods."
    AuthMethodsResponse:
      type: object
      required: [email, methods]
      properties:
        email: { type: string, format: email }
        tenant_id:
          type: string
          description: "The resolved user's tenant id, or empty string when the email matches no active user. Used to build the SSO authorize URL."
        methods:
          type: array
          items: { $ref: "#/components/schemas/AuthMethod" }
    # --- Member invitations ------------------------------------------
...
    GroupRoleMapping:
      type: object
      additionalProperties: false
      description: "Maps an external IdP group to a tenant role (api.GroupRoleMapping)."
      required:
        - external_group_name
        - role_id
      properties:
        id:
          type: string
        sso_provider_id:
          type: string
        external_group_name:
          type: string
        role_id:
          type: string
        role_name:
          type: string
    SSOProvider:
      type: object
      additionalProperties: false
      description: >
        An SSO provider (api.SSOProvider). The base fields are always present;
        the OAuth/OIDC + SAML config fields and group mappings are omitempty and
        appear only in the admin (full-config) view. Client secrets are never
        returned. attribute_mapping serializes as JSON null when unset.
      required:
        - id
        - tenant_id
        - provider_type
        - provider_name
        - is_enabled
        - is_default
        - auto_provision_users
        - attribute_mapping
        - created_at
        - updated_at
      properties:
        id:
          type: string
        tenant_id:
          type: string
        provider_type:
          type: string
        provider_name:
          type: string
        is_enabled:
          type: boolean
        is_default:
          type: boolean
        auto_provision_users:
          type: boolean
        attribute_mapping:
          type: [object, "null"]
        default_role_id:
          type: string
        allowed_domains:
          type: array
          items:
            type: string
        groups_claim_name:
          type: string
        group_role_mappings:
          type: array
          items:
            $ref: "#/components/schemas/GroupRoleMapping"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        client_id:
          type: string
        auth_url:
          type: string
        token_url:
          type: string
        userinfo_url:
          type: string
        scopes:
          type: string
        saml_entity_id:
          type: string
        saml_sso_url:
          type: string
    SSOProvidersResponse:
      type: object
      additionalProperties: false
      description: 'GET /tenant/sso/providers — `{ "providers": [...] }`.'
      required:
        - providers
      properties:
        providers:
          type: array
          items:
            $ref: "#/components/schemas/SSOProvider"
    SSOProviderResponse:
      type: object
      additionalProperties: false
      description: 'POST /tenant/sso/providers (201) — `{ "provider": {...} }`.'
      required:
        - provider
      properties:
        provider:
          $ref: "#/components/schemas/SSOProvider"
    CreateSSOProviderRequest:
      type: object
      additionalProperties: false
      description: >
        Request body for POST /tenant/sso/providers (api.CreateSSOProviderRequest).
        provider_type + provider_name are required; the rest depend on the type
        (OAuth2 vs SAML). client_secret / saml_private_key are write-only and are
        encrypted at rest.
      required:
        - provider_type
        - provider_name
      properties:
        provider_type:
          type: string
        provider_name:
          type: string
        is_enabled:
          type: boolean
        is_default:
          type: boolean
        auto_provision_users:
          type: boolean
        attribute_mapping:
          type: object
        default_role_id:
          type: string
        allowed_domains:
          type: array
          items:
            type: string
        groups_claim_name:
          type: string
        group_role_mappings:
          type: array
          items:
            $ref: "#/components/schemas/GroupRoleMapping"
        client_id:
          type: string
        client_secret:
          type: string
        auth_url:
          type: string
        token_url:
          type: string
        userinfo_url:
          type: string
        scopes:
          type: string
        saml_entity_id:
          type: string
        saml_sso_url:
          type: string
        saml_certificate:
          type: string
        saml_private_key:
          type: string
    UpdateSSOProviderRequest:
      type: object
      additionalProperties: false
      description: >
        Request body for PUT /tenant/sso/providers/{id}
        (api.UpdateSSOProviderRequest). All fields optional — only provided fields
        are updated. At least one field (or group_role_mappings) must be present.
        client_secret / saml_private_key are write-only and re-encrypted at rest.
      properties:
        provider_name:
          type: string
        is_enabled:
          type: boolean
        is_default:
          type: boolean
        auto_provision_users:
          type: boolean
        attribute_mapping:
          type: object
        default_role_id:
          type: string
        allowed_domains:
          type: array
          items:
            type: string
        groups_claim_name:
          type: string
        group_role_mappings:
          type: array
          items:
            $ref: "#/components/schemas/GroupRoleMapping"
        client_id:
          type: string
        client_secret:
          type: string
        auth_url:
          type: string
        token_url:
          type: string
        userinfo_url:
          type: string
        scopes:
          type: string
        saml_entity_id:
          type: string
        saml_sso_url:
          type: string
        saml_certificate:
          type: string
        saml_private_key:
          type: string
    SSOTestResult:
      type: object
      additionalProperties: true
      description: >
        POST /tenant/sso/providers/{id}/test — config-validation result. status is
        "success" or "error"; message explains. (Open object — additional
        diagnostic keys may be added.)
      required:
        - provider_id
        - provider_type
        - status
        - message
      properties:
        provider_id:
          type: string
        provider_type:
          type: string
        status:
          type: string
        message:
          type: string
...
    LegacyError:
      type: object
      additionalProperties: true
      description: >
        CURRENT error shape returned by these endpoints today: a single
        human-readable string under `error`. Some 400s carry extra sibling keys
        (e.g. `details`, `required_permission`), so additionalProperties is
        permitted here. Unstructured — cannot be branched on safely. Superseded
        by the ADR-0002 `Error` envelope below as endpoints are hardened.
      required: [error]
      properties:
        error:
          type: string
...
    MessageResponse:
      type: object
      additionalProperties: false
      description: Generic mutation acknowledgement (a single human-readable string under `message`).
      required: [message]
      properties:
        message:
          type: string
...